working on collisions

+++ Source/MonoGame.Extended/Maps/Tiled/CollisionWorldExtensions.cs
@@ -8,7 +8,7 @@ namespace MonoGame.Extended.Maps.Tiled
         public static CollisionGrid CreateGrid(this CollisionWorld world, TiledTileLayer tileLayer)
         {
             var data = tileLayer.Tiles
-                .Select(t => (byte)t.Id)
+                .Select(t => t.Id)
                 .ToArray();
 
             return world.CreateGrid(data, tileLayer.Width, tileLayer.Height, tileLayer.TileWidth, tileLayer.TileHeight);
